修改格式，便于作为依赖仓库进行调用
