+++ utils/file.go
@@ -1,0 +1,31 @@
+package utils
+
+import (
+	"github.com/LubyRuffy/chrome_proxy/models"
+	"os"
+)
+
+// WriteTempFile 写入临时文件
+// 如果writeF是nil，就只返回生成的一个临时空文件路径
+// 返回文件名和错误
+func WriteTempFile(ext string, writeF func(f *os.File) error) (fn string, err error) {
+	var f *os.File
+	if len(ext) > 0 {
+		ext = "*" + ext
+	}
+	f, err = os.CreateTemp(os.TempDir(), models.DefaultTmpFilePrefix+ext)
+	if err != nil {
+		return
+	}
+	defer f.Close()
+
+	fn = f.Name()
+
+	if writeF != nil {
+		err = writeF(f)
+		if err != nil {
+			return
+		}
+	}
+	return
+}
